+++ threadify/__init__.py
@@ -4,7 +4,6 @@ from threadify.client import (
 )
 from threadify.connection import Connection
 from threadify.data_retriever import ArchivedStep, ArchivedThread, DataRetriever
-from threadify.management import EntityProfileManager, ManagementAPIError, profile_slug
 from threadify.models import (
     AccessLevel,
     CompleteDataOptions,
@@ -17,14 +16,19 @@ from threadify.models import (
     StepResult,
     SubStepData,
     ThreadEndResponse,
+    ThreadOptions,
     WaitOptions,
 )
 from threadify.notification import Notification
 from threadify.otel_exporter import ThreadifySpanExporter
 from threadify.step import DuplicateStepError, ThreadStep, is_duplicate_error
 from threadify.thread import ThreadInstance
+from threadify.waiting import PermissionGrant, ThreadifyError, WaitResult
 
 __all__ = [
+    "PermissionGrant",
+    "ThreadifyError",
+    "WaitResult",
     "Threadify",
     "ThreadifyFactory",
     "Connection",
@@ -34,9 +38,6 @@ __all__ = [
     "is_duplicate_error",
     "Notification",
     "DataRetriever",
-    "EntityProfileManager",
-    "ManagementAPIError",
-    "profile_slug",
     "ArchivedThread",
     "ArchivedStep",
     "ThreadifySpanExporter",
@@ -46,6 +47,7 @@ __all__ = [
     "InviteOptions",
     "InviteResponse",
     "ThreadEndResponse",
+    "ThreadOptions",
     "WaitOptions",
     "NotificationData",
     "RefQuery",
@@ -54,4 +56,4 @@ __all__ = [
     "AccessLevel",
 ]
 
-__version__ = "0.1.0"
+__version__ = "0.3.1"